+++ ChangeLog.rst
@@ -2,6 +2,15 @@
  Pareto's ChangeLog
 ====================
 
+0.1.0 (2026-04-29)
+==================
+
+* Updated API calls to be compatible with the current version of Lem:
+
+  - ``lem:backward-up-list`` now requires a point argument
+  - ``lem:region-beginning`` and ``lem:region-end`` now require a buffer argument
+  - Replaced ``lem:mark-set`` with ``lem:set-cursor-mark``
+
 0.1.0 (2020-01-12)
 ==================